Split the extention into several modules
- also resolves #8
