+++ core/deployment/src/test/java/io/quarkiverse/businessscore/test/BusinessScoreTest.java
