+++ core/deployment/src/test/java/io/quarkiverse/businessscore/test/MyBusinessService.java
