+++ core/runtime/src/main/java/io/quarkiverse/businessscore/Scored.java
