+++ core/runtime/src/main/java/io/quarkiverse/businessscore/runtime/ScoredInterceptor.java
