+++ core/runtime/src/main/resources/META-INF/quarkus-extension.yaml
