+++ docs/modules/ROOT/pages/includes/quarkus-business-score-http.adoc
@@ -1,0 +1,31 @@
+
+:summaryTableId: quarkus-business-score-http
+[.configuration-legend]
+icon:lock[title=Fixed at build time] Configuration property fixed at build time - All other configuration properties are overridable at runtime
+[.configuration-reference.searchable, cols="80,.^10,.^10"]
+|===
+
+h|[[quarkus-business-score-http_configuration]]link:#quarkus-business-score-http_configuration[Configuration property]
+
+h|Type
+h|Default
+
+a|icon:lock[title=Fixed at build time] [[quarkus-business-score-http_quarkus-business-score-http-root-path]]`link:#quarkus-business-score-http_quarkus-business-score-http-root-path[quarkus.business-score.http.root-path]`
+
+
+[.description]
+--
+The root path for management endpoints.
+
+By default, this value is resolved as a path relative to `quarkus.http.non-application-root-path`. If the management interface is enabled, the value will be resolved as a path relative to `quarkus.management.root-path`.
+
+ifdef::add-copy-button-to-env-var[]
+Environment variable: env_var_with_copy_button:+++QUARKUS_BUSINESS_SCORE_HTTP_ROOT_PATH+++[]
+endif::add-copy-button-to-env-var[]
+ifndef::add-copy-button-to-env-var[]
+Environment variable: `+++QUARKUS_BUSINESS_SCORE_HTTP_ROOT_PATH+++`
+endif::add-copy-button-to-env-var[]
+--|string 
+|`business-score`
+
+|===
