+++ docs/modules/ROOT/pages/includes/quarkus-business-score-scheduler.adoc
@@ -1,0 +1,31 @@
+
+:summaryTableId: quarkus-business-score-scheduler
+[.configuration-legend]
+icon:lock[title=Fixed at build time] Configuration property fixed at build time - All other configuration properties are overridable at runtime
+[.configuration-reference.searchable, cols="80,.^10,.^10"]
+|===
+
+h|[[quarkus-business-score-scheduler_configuration]]link:#quarkus-business-score-scheduler_configuration[Configuration property]
+
+h|Type
+h|Default
+
+a|icon:lock[title=Fixed at build time] [[quarkus-business-score-scheduler_quarkus-business-score-scheduler-cron]]`link:#quarkus-business-score-scheduler_quarkus-business-score-scheduler-cron[quarkus.business-score.scheduler.cron]`
+
+
+[.description]
+--
+The CRON trigger used to schedule an automatic self-test.
+
+If `BusinessScore.ZombieStatus++#++isZombie()` returns `true` then a warning message is logged and a CDI event of type `BusinessScore.ZombieStatus` is fired synchronously.
+
+ifdef::add-copy-button-to-env-var[]
+Environment variable: env_var_with_copy_button:+++QUARKUS_BUSINESS_SCORE_SCHEDULER_CRON+++[]
+endif::add-copy-button-to-env-var[]
+ifndef::add-copy-button-to-env-var[]
+Environment variable: `+++QUARKUS_BUSINESS_SCORE_SCHEDULER_CRON+++`
+endif::add-copy-button-to-env-var[]
+--|string 
+|`0 0 0/6 * * ?`
+
+|===
