+++ docs/modules/ROOT/pages/index.adoc
@@ -6,9 +6,18 @@ This extension offers a convenient way to detect application _zombies_, i.e. app
 The "business value" means anything that helps to solve a problem of an application user. 
 Infrastructure stuff (logs, stats, metrics, health checks) usually does not produce a business value.
 
-== Installation
+== Key concepts
 
-If you want to use this extension, you need to add the `io.quarkiverse.businessscore:quarkus-business-score` extension first to your build file.
+In a nutshell, this extension manages the **business score** which is a numeric representation of the business value. 
+An application emits **score records** to increase the business score, either directly with the `BusinessScore` API or with an interceptor binding.
+The records are only kept for the specific amount of time - **time window**. 
+The sum of all score records in the current time window is the **current business score**. 
+The **zombie threshold** defines the point at which the application is considered a _zombie_.
+An external system or the application itself can periodically check the business score and react appropriately.
+
+== Core
+
+If you want to use this extension, you need to add the `io.quarkiverse.businessscore:quarkus-business-score` extension first.
 
 For instance, with Maven, add the following dependency to your POM file:
 
@@ -21,16 +30,7 @@ For instance, with Maven, add the following dependency to your POM file:
 </dependency>
 ----
 
-== The concepts
-
-In a nutshell, the extension manages the **business score** which is a numeric representation of the business value. 
-An application emits **score records** to increase the business score, either directly with the `BusinessScore` API or with an interceptor binding.
-The records are only kept for the specific amount of time - **time window**. 
-The sum of all score records in the current time window is the **current business score**. 
-The **zombie threshold** defines the point at which the application is considered a _zombie_.
-An external system or the application itself can periodically check the business score and react appropriately.
-
-== How to score
+=== How to score
 
 Quarkus provides a CDI bean of type `io.quarkiverse.businessscore.BusinessScore` that can be used to increase and also to query the current score.
 
@@ -76,7 +76,7 @@ public class MyService {
 ----
 <1> Increase the business score of the current application by `10`.
 
-== How to check the business score
+=== How to check the business score
 
 You can inject the `io.quarkiverse.businessscore.BusinessScore` and check the "zombie" status with the `BusinessScore#test()` method.
 
@@ -102,8 +102,73 @@ public class MyService {
 ----
 <1> Inject the business score of the current application.
 
-However, this extension tests the current application periodically and if a a _zombie_ is detected then it logs a warning and fires a CDI event of type `io.quarkiverse.businessscore.BusinessScore.ZombieStatus` synchronously.
-The interval is configurable with the `quarkus.business-score.zombie-check-interval` config property.
+== HTTP module
+
+This module integrates with the https://quarkus.io/guides/http-reference[HTTP extension] and registers two additional endpoints:
+
+- `/q/business-score` - shows the current score, zombie status, etc.
+- `/q/business-score/records` - shows all score records in the current time window
+ 
+
+[source,xml,subs=attributes+]
+----
+<dependency>
+    <groupId>io.quarkiverse.businessscore</groupId>
+    <artifactId>quarkus-business-score-http</artifactId>
+    <version>{project-version}</version>
+</dependency>
+----
+
+.`/q/business-score` response example
+[source,json]
+----
+{
+    "serverTime": "2024-06-13T15:14:20.794528118",
+    "application": {
+        "name": "quarkus-business-score-integration-tests",
+        "version": "999-SNAPSHOT"
+    },
+    "zombie": false,
+    "score": 1,
+    "zombieThreshold": 20,
+    "initialZombieThreshold": 10,
+    "timeWindow": "PT72H",
+    "initialTimeWindow": "PT168H"
+}
+----
+
+.`/q/business-score/records` response example
+[source,json]
+----
+{
+    "serverTime": "2024-06-13T15:14:21.259640872",
+    "application": {
+        "name": "quarkus-business-score-integration-tests",
+        "version": "999-SNAPSHOT"
+    },
+    "records": [
+        {
+            "value": 1,
+            "timestamp": "2024-06-13T13:14:20.735Z"
+        }
+    ]
+}
+----
+
+== Scheduler module
+
+This module integrates with the https://quarkus.io/guides/scheduler-reference[Scheduler] extension and registers an additional `@Scheduled` method.
+I.e. the current application is tested periodically and if a _zombie_ is detected then a warning message is logged and a CDI event of type `io.quarkiverse.businessscore.BusinessScore.ZombieStatus` is fired synchronously.
+The CRON trigger used to schedule an automatic self-test can be configured with the `quarkus.business-score.scheduler.cron` configuration property.
+
+[source,xml,subs=attributes+]
+----
+<dependency>
+    <groupId>io.quarkiverse.businessscore</groupId>
+    <artifactId>quarkus-business-score-scheduler</artifactId>
+    <version>{project-version}</version>
+</dependency>
+----
 
 .CDI observer example
 [source,java]
@@ -120,9 +185,28 @@ public class MyService {
 
 }
 ----
-<1> Observe the `ZombieStatus`.
+<1> Observe the `ZombieStatus` and react appropriately.
+
+== Health module
+
+This module integrates with the https://quarkus.io/guides/smallrye-health[SmallRye Health] extension and registers a https://smallrye.io/docs/smallrye-health/3.0.1/wellness.html[`@Wellness` health check] automatically.
+The health check is accessible through the `/q/health/well` endpoint.
+ 
+[source,xml,subs=attributes+]
+----
+<dependency>
+    <groupId>io.quarkiverse.businessscore</groupId>
+    <artifactId>quarkus-business-score-health</artifactId>
+    <version>{project-version}</version>
+</dependency>
+----
 
 [[extension-configuration-reference]]
-== Extension Configuration Reference
+=== Extension Configuration Reference
 
 include::includes/quarkus-business-score.adoc[leveloffset=+1, opts=optional]
+
+include::includes/quarkus-business-score-http.adoc[leveloffset=+1, opts=optional]
+
+include::includes/quarkus-business-score-scheduler.adoc[leveloffset=+1, opts=optional]
+
