+++ docs/pom.xml
@@ -18,6 +18,16 @@
             <artifactId>quarkus-business-score-deployment</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>io.quarkiverse.businessscore</groupId>
+            <artifactId>quarkus-business-score-http-deployment</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkiverse.businessscore</groupId>
+            <artifactId>quarkus-business-score-scheduler-deployment</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
@@ -66,6 +76,16 @@
                                 <resource>
                                     <directory>${project.basedir}/../target/asciidoc/generated/config/</directory>
                                     <include>quarkus-business-score.adoc</include>
+                                    <filtering>false</filtering>
+                                </resource>
+                                <resource>
+                                    <directory>${project.basedir}/../target/asciidoc/generated/config/</directory>
+                                    <include>quarkus-business-score-http.adoc</include>
+                                    <filtering>false</filtering>
+                                </resource>
+                                <resource>
+                                    <directory>${project.basedir}/../target/asciidoc/generated/config/</directory>
+                                    <include>quarkus-business-score-scheduler.adoc</include>
                                     <filtering>false</filtering>
                                 </resource>
                                 <resource>
