+++ http/runtime/src/main/java/io/quarkiverse/businessscore/http/runtime/BusinessScoreHttp.java
@@ -1,0 +1,22 @@
+package io.quarkiverse.businessscore.http.runtime;
+
+import jakarta.inject.Inject;
+import jakarta.inject.Singleton;
+
+import io.quarkiverse.businessscore.BusinessScore;
+import io.quarkiverse.businessscore.runtime.BusinessScoreConfig;
+import io.quarkus.runtime.ApplicationConfig;
+
+@Singleton
+public class BusinessScoreHttp {
+
+    @Inject
+    BusinessScore score;
+
+    @Inject
+    ApplicationConfig appConfig;
+
+    @Inject
+    BusinessScoreConfig businessScoreConfig;
+
+}
