+++ pom.xml
@@ -12,8 +12,11 @@
   <packaging>pom</packaging>
   <name>Quarkus Business Score - Parent</name>
   <modules>
-    <module>deployment</module>
-    <module>runtime</module>
+    <module>core</module>
+    <module>http</module>
+    <module>health</module>
+    <module>scheduler</module>
+    <module>integration-tests</module>
     <module>docs</module>
   </modules>
   <scm>
