+++ scheduler/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,9 @@
+name: Business Score - Scheduler
+#description: Do something useful.
+metadata:
+#  keywords:
+#    - business-score
+#  guide: https://quarkiverse.github.io/quarkiverse-docs/business-score/dev/ # To create and publish this guide, see https://github.com/quarkiverse/quarkiverse/wiki#documenting-your-extension
+#  categories:
+#    - "miscellaneous"
+#  status: "preview"
